+++ examples/java-jfr/rideshare/load-generator.py
@@ -4,9 +4,9 @@ import time
 import traceback
 
 HOSTS = [
-    'us-east-1',
-    'us-west-1',
-    'eu-west-1',
+    'us-east',
+    'eu-north',
+    'ap-south',
 ]
 
 VEHICLES = [