+++ examples/java-jfr/rideshare/src/main/java/org/example/rideshare/OrderService.java
@@ -15,20 +15,7 @@ public class OrderService {
 
     public static final Duration OP_DURATION = Duration.of(200, ChronoUnit.MILLIS);
 
-    public void orderBike(int searchRadius) {
-        findNearestVehicle(searchRadius, "bike");
-    }
-
-    public void orderScooter(int searchRadius) {
-        findNearestVehicle(searchRadius, "scooter");
-    }
-
-    public void orderCar(int searchRadius) {
-        findNearestVehicle(searchRadius, "car");
-    }
-
-
-    private synchronized void findNearestVehicle(int searchRadius, String vehicle) {
+    public synchronized void findNearestVehicle(int searchRadius, String vehicle) {
         Pyroscope.LabelsWrapper.run(new LabelsSet("vehicle", vehicle), () -> {
             AtomicLong i = new AtomicLong();
             Instant end = Instant.now()
@@ -50,11 +37,11 @@ public class OrderService {
         while (Instant.now().compareTo(end) <= 0) {
             i.incrementAndGet();
         }
-        // Every other minute this will artificially create make requests in us-west-1 region slow
+        // Every other minute this will artificially create make requests in eu-north region slow
         // this is just for demonstration purposes to show how performance impacts show up in the
         // flamegraph
         boolean force_mutex_lock = Instant.now().atZone(ZoneOffset.UTC).getMinute() % 2 == 0;
-        if (System.getenv("REGION").equals("us-west-1") && force_mutex_lock) {
+        if (System.getenv("REGION").equals("eu-north") && force_mutex_lock) {
             mutexLock(searchRadius);
         }
     }